requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-ollama
langchain-groq
langsmith
langserve
python-dotenv
requests
fastapi
uvicorn
sse_starlette
pypdf
sentence-transformers
faiss-cpu